Refactorizar datasets de enemigos. Se modula grupos de enemies por ui map y ui algorithm, del mismo modo la data de behaviors y paths queda separada para mejor integracion y reutilizacion. Se ajusta a nivel general el nuevo modelo de especificacion de entidades.
